refactor(data): unify loaders with lazy Supabase + optional cache; deprecate legacy fetchers
