--- tests/test_fetch_data_async.py
@@ -1,125 +1,0 @@
-"""Tests for asynchronous Supabase pagination utilities."""
-
-import os
-import sys
-
-import httpx
-import pandas as pd
-import pytest
-
-sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
-from cointrainer.data.loader import (
-    fetch_data_async,
-    fetch_data_between_async,
-    fetch_data_range_async,
-    fetch_table_async,
-)
-
-# Common mock transport used across tests
-PAGES = [
-    [{"id": 1, "val": 10}, {"id": 2, "val": 20}],
-    [{"id": 3, "val": 30}, {"id": 4, "val": 40}],
-    [{"id": 5, "val": 50}],
-]
-CHUNK_SIZE = 2
-
-
-def make_transport():
-    """Return an ``httpx.MockTransport`` serving ``PAGES``."""
-
-    def handler(request: httpx.Request) -> httpx.Response:
-        if request.headers.get("Range"):
-            start = int(request.headers["Range"].split("-", 1)[0])
-            idx = start // CHUNK_SIZE
-        else:
-            offset = int(request.url.params.get("offset", "0"))
-            idx = offset // CHUNK_SIZE
-        data = PAGES[idx] if idx < len(PAGES) else []
-        return httpx.Response(200, json=data)
-
-    return httpx.MockTransport(handler)
-
-
-@pytest.mark.asyncio
-async def test_fetch_data_range_async(monkeypatch):
-    """``fetch_data_range_async`` should concatenate all paginated chunks."""
-    transport = make_transport()
-    real_client = httpx.AsyncClient
-
-    def fake_client(**kwargs):
-        kwargs.setdefault("transport", transport)
-        kwargs.setdefault("base_url", "https://sb.example.com")
-        return real_client(**kwargs)
-
-    monkeypatch.setattr(httpx, "AsyncClient", fake_client)
-    monkeypatch.setenv("SUPABASE_URL", "https://sb.example.com")
-    monkeypatch.setenv("SUPABASE_KEY", "test")
-
-    async with fake_client() as client:
-        df1 = await fetch_table_async("ohlc_data", page_size=CHUNK_SIZE, client=client)
-        df1 = await fetch_table_async("ohlc_data", page_size=CHUNK_SIZE, client=client)
-
-    df2 = await fetch_data_range_async(
-        "ohlc_data", "start", "end", chunk_size=CHUNK_SIZE
-    )
-    df3 = await fetch_data_between_async(
-        "ohlc_data", "start", "end", chunk_size=CHUNK_SIZE
-    )
-
-    expected = pd.concat([pd.DataFrame(p) for p in PAGES], ignore_index=True)
-    pd.testing.assert_frame_equal(df1, expected)
-    pd.testing.assert_frame_equal(df2, expected)
-    pd.testing.assert_frame_equal(df3, expected)
-    df = await fetch_data_range_async(
-        "ohlc_data",
-        "2021-01-01",
-        "2021-01-02",
-        chunk_size=CHUNK_SIZE,
-    )
-
-    expected = pd.concat([pd.DataFrame(p) for p in PAGES], ignore_index=True)
-    pd.testing.assert_frame_equal(df, expected)
-
-
-@pytest.mark.asyncio
-async def test_fetch_data_async_table(monkeypatch):
-    """``fetch_data_async`` should paginate the whole table."""
-    transport = make_transport()
-    real_client = httpx.AsyncClient
-
-    def fake_client(**kwargs):
-        kwargs.setdefault("transport", transport)
-        kwargs.setdefault("base_url", "https://sb.example.com")
-        return real_client(**kwargs)
-
-    monkeypatch.setattr(httpx, "AsyncClient", fake_client)
-    monkeypatch.setenv("SUPABASE_URL", "https://sb.example.com")
-    monkeypatch.setenv("SUPABASE_KEY", "test")
-
-    async with fake_client() as client:
-        df = await fetch_data_async("ohlc_data", page_size=CHUNK_SIZE, client=client)
-
-    expected = pd.concat([pd.DataFrame(p) for p in PAGES], ignore_index=True)
-    pd.testing.assert_frame_equal(df, expected)
-
-
-@pytest.mark.asyncio
-async def test_async_headers_use_jwt(monkeypatch):
-    transport = make_transport()
-    real_client = httpx.AsyncClient
-    captured = {}
-
-    def fake_client(**kwargs):
-        captured["headers"] = kwargs.get("headers")
-        kwargs.setdefault("transport", transport)
-        kwargs.setdefault("base_url", "https://sb.example.com")
-        return real_client(**kwargs)
-
-    monkeypatch.setattr(httpx, "AsyncClient", fake_client)
-    monkeypatch.setenv("SUPABASE_URL", "https://sb.example.com")
-    monkeypatch.setenv("SUPABASE_KEY", "test")
-    monkeypatch.setenv("SUPABASE_JWT", "token123")
-
-    await fetch_data_range_async("ohlc_data", "s", "e")
-
-    assert captured["headers"]["Authorization"] == "Bearer token123"
